==> Code/01_Data_Collection.ipynb <==
import requests
import calendar
import time
import pandas as pd

def get_posts(num_posts, req_params, before):
    """Get specified number of subreddit posts using Pushshift API
    
    This function will request the number of posts at a time specified
    in the arguments (max 100), up to the total number of posts also 
    specified, with a configurable timeout between requests.  It will
    start at the provided epoch time and work backwards.  Also, it will
    end and return the posts before the specified number of posts is
    reached if the request fails more than 10 times in a row, or 0 
    posts is returned in a request.
    
    Args:
        num_posts (int): Total number of posts desired
        req_params (dict): Dictionary of request parameters
            key (string): url, value (string): Base API url
            key (string): subreddit, value (string): subreddit name
            key (string): size, value (int): number of posts per request
            key (string): fields, value (list): list of subreddit fields
            key (string): secs, value (int): seconds between requests
        before (int): epoch time to start requests from
        
    Return:
        Dataframe of results
    """
    # instantiate list for storing all posts
    all_posts = []
    
    # make requests until required number of posts is reached
    while(len(all_posts) < num_posts):
        params = {
            'subreddit': req_params['subreddit'],
            'size': req_params['size'],
            'fields': req_params['fields'],
            'before': before
        }
        res = requests.get(req_params['url'], params)
        
        # counter for failed requests
        # resets with a successful request
        # breaks loop if unsuccessful requests is greater than 10
        num_failed = 0
        if res.status_code == 200:
            num_failed = 0
            posts = res.json()['data']
            if len(posts) > 0:
                all_posts.append(posts)
                before = posts[-1]['created_utc']
            else:
                print('No more posts')
                break
        elif num_failed > 10:
            print('requests unsuccessful')
            break
        else:
            num_failed += 1
            
        # wait time between requests
        time.sleep(req_params['secs'])
        
    print('Complete')    
    return pd.DataFrame(all_posts)

## Fitness

#instantiate dataframe before first run only
#fitness_df = pd.DataFrame()

#read in existing data to start where left off
fitness_df = pd.read_csv('../Data/fitness.csv', index_col=0)
fitness_df.head(2)

#parameters for request
req_params = {
    'url': 'https://api.pushshift.io/reddit/search/submission',
    'subreddit': 'Fitness',
    'size': 100,
    'fields': ['subreddit','author','title','selftext','num_comments','num_crossposts','score','created_utc'],
    'secs': 10
}

#current time for first run
now = calendar.timegm(time.gmtime())

#time for starting from where left off
before = fitness_df.iloc[-1]['created_utc']

fitness_df = pd.concat([fitness_df, get_posts(4000, req_params, before)])

len(fitness_df)

fitness_df.head()

fitness_df.info()

#save to file
fitness_df.to_csv('../Data/fitness.csv')

## Bodyweight Fitness

#instantiate dataframe before first run only
#bodyweight_df = pd.DataFrame()

#read in existing data to start where left off
bodyweight_df = pd.read_csv('../Data/bodyweight.csv', index_col=0)
bodyweight_df.head(2)

#parameters for request
req_params = {
    'url': 'https://api.pushshift.io/reddit/search/submission',
    'subreddit': 'bodyweightfitness',
    'size': 100,
    'fields': ['subreddit','author','title','selftext','num_comments','num_crossposts','score','created_utc'],
    'secs': 10
}

#current time for first run
now = calendar.timegm(time.gmtime())

#time for starting from where left off
before = bodyweight_df.iloc[-1]['created_utc']

bodyweight_df = pd.concat([bodyweight_df, get_posts(9500, req_params, before)])

len(bodyweight_df)

bodyweight_df.head()

bodyweight_df.info()

#save to file
bodyweight_df.to_csv('../Data/bodyweight.csv')
