==> Code/02_Data_Cleaning.ipynb <==
import pandas as pd
import numpy as np

# Cleaning r/Bodyweight Fitness

bodyweight_df = pd.read_csv('../Data/bodyweight.csv', index_col=0)

bodyweight_df.head(2)

bodyweight_df.info()

# drop nulls
bodyweight_df.dropna(inplace=True)

# drop rows where post is removed or deleted
bodyweight_df = bodyweight_df.loc[bodyweight_df['selftext'] != '[removed]']
bodyweight_df = bodyweight_df.loc[bodyweight_df['selftext'] != '[deleted]']

bodyweight_df.info()

# drop rows containing the subreddit name in the post
bodyweight_df = bodyweight_df[~bodyweight_df['selftext'].str.contains(f"r/{bodyweight_df.iloc[0,6]}")]

bodyweight_df.info()

# drop rows where post is shorter than title
# this gets rid of posts that were just media with little collateral
bodyweight_df = bodyweight_df[bodyweight_df['selftext'].str.len() >= bodyweight_df['title'].str.len()]

bodyweight_df.info()

#saving cleaned data
bodyweight_df.to_csv('../Data/bodyweight_clean.csv')

# function with all the steps used to clean r/bodyweightfitness
def clean_data(df):
    """Remove unwanted rows from a dataframe of subreddit posts.
    
    For this function, the dataframe must have a column named
    'selftext' with the text of the subreddit post.  It must
    also contain the name of the subreddit in column of index 6.
    This function will drop rows with the following conditions:
    - Rows with null values
    - Rows with removed or deleted posts
    - Rows containing the subreddit name in the post
    - Rows where the post is shorter than the title
    
    args:
        df (pandas dataframe): dataframe to clean
        
    returns:
        pandas dataframe: cleaned dataframe
    """
    # drop nulls
    df.dropna(inplace=True)
    
    # drop rows with removed or deleted posts
    df = df.loc[df['selftext'] != '[removed]']
    df = df.loc[df['selftext'] != '[deleted]']
    
    # drop rows containing the subreddit name in the post
    df = df[~df['selftext'].str.contains(f"r/{df.iloc[0,6]}")]
    
    # drop rows where post is shorter than title
    df = df[df['selftext'].str.len() >= df['title'].str.len()]
    
    # reset the index
    df.reset_index(drop=True)
    
    return df

# Cleaning r/Fitness

fitness_df = pd.read_csv('../Data/fitness.csv', index_col=0)

fitness_df.head(2)

fitness_df.info()

# apply same cleaning as r/bodyweightfitness using function
fitness_df_clean = clean_data(fitness_df)

fitness_df_clean.info()

# saving cleaned data
fitness_df_clean.to_csv('../Data/fitness_clean.csv')
