==> Code/04a_Model_Selection.ipynb <==
import pandas as pd
import numpy as np
import warnings

from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.preprocessing import StandardScaler
from sklearn.pipeline import make_pipeline, Pipeline
from sklearn.model_selection import train_test_split, GridSearchCV

from sklearn.linear_model import LogisticRegression
from sklearn.neighbors import KNeighborsClassifier
from sklearn.naive_bayes import MultinomialNB
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.svm import SVC

#Ignoring warnings about memory from running gridsearch with n_jobs=-1
warnings.filterwarnings('ignore')

fitness_df = pd.read_csv('../Data/fitness_clean.csv', index_col=0)
bodyweight_df = pd.read_csv('../Data/bodyweight_clean.csv', index_col=0)

all_posts = pd.concat([fitness_df, bodyweight_df], ignore_index=True)

# null model accuracy
all_posts['subreddit'].value_counts(normalize=True)

X = all_posts['selftext']
y = all_posts['subreddit']

X_train, X_test, y_train, y_test = train_test_split(X, y)

## Logistic Regression

pipe = make_pipeline(TfidfVectorizer(stop_words='english'), StandardScaler(with_mean=False), LogisticRegression(max_iter=10_000))
params = {
    'tfidfvectorizer__max_features': [800, 1000],
    'tfidfvectorizer__ngram_range': [(1,1), (1,2)],
    'logisticregression__C': [0.001, 0.01, 0.1]
}
grid = GridSearchCV(pipe, params, n_jobs=-1)

grid.fit(X_train, y_train)

grid.score(X_train, y_train)

grid.score(X_test, y_test)

grid.best_params_

# KNN

pipe = make_pipeline(TfidfVectorizer(stop_words='english'), StandardScaler(with_mean=False), KNeighborsClassifier())
params = {
    'tfidfvectorizer__max_features': [800, 1000],
    'tfidfvectorizer__ngram_range': [(1,1), (1,2)],
    'kneighborsclassifier__n_neighbors': [3, 5, 15]
}
grid = GridSearchCV(pipe, params, n_jobs=-1)

grid.fit(X_train, y_train)

grid.score(X_train, y_train)

grid.score(X_test, y_test)

# GridSearch Multiple Estimators

pipe = Pipeline([
    ('vect', TfidfVectorizer(stop_words='english', max_features=1000)),
    ('scaler', StandardScaler(with_mean=False)),
    ('clf', MultinomialNB())
])

params = [
    {
        'clf': [MultinomialNB()]
    }, {
        'clf': [DecisionTreeClassifier()]
    }, {
        'clf': [RandomForestClassifier()]
    }, {
        'clf': [SVC()]
    }
]

grid = GridSearchCV(pipe, params, n_jobs=-1)

grid.fit(X_train, y_train)

grid.score(X_train, y_train)

grid.score(X_test, y_test)

pd.DataFrame(grid.cv_results_)

## Conclusions
Since the task is to predict which of two subreddits a given post belongs to with no preference for either, I will use basic accuracy as my primary metric.  In addition, the classes are well balanced which also lends itself to using accuracy as the main metric.  Based on the accuracy scores, I will focus on LogisticRegression, MultinomialNB, RandomForestClassifier, and SVC.
