==> Code/04b_Bayes_Models.ipynb <==
import pandas as pd
import warnings
import re

from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.preprocessing import StandardScaler
from sklearn.pipeline import make_pipeline
from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.naive_bayes import MultinomialNB

from nltk.tokenize import word_tokenize
from nltk.stem import WordNetLemmatizer, PorterStemmer
from nltk.corpus import stopwords

#Ignoring warnings about memory from running gridsearch with n_jobs=-1
warnings.filterwarnings('ignore')

fitness_df = pd.read_csv('../Data/fitness_clean.csv', index_col=0)
bodyweight_df = pd.read_csv('../Data/bodyweight_clean.csv', index_col=0)

all_posts = pd.concat([fitness_df, bodyweight_df], ignore_index=True)

# Baseline
all_posts['subreddit'].value_counts(normalize=True)

X = all_posts['selftext']
y = all_posts['subreddit']

X_train, X_test, y_train, y_test = train_test_split(X, y)

## Grid Search Basic MultinomialNB

pipe = make_pipeline(TfidfVectorizer(), StandardScaler(with_mean=False), MultinomialNB())
params = {
    'tfidfvectorizer__max_features': [None, 400, 800, 1000],
    'tfidfvectorizer__ngram_range': [(1,1), (1,2)],
    'tfidfvectorizer__stop_words': [None, 'english'],
    'multinomialnb__alpha': [0.001, 0.01, 1]
}
grid = GridSearchCV(pipe, params, n_jobs=-1)

grid.fit(X_train, y_train)

grid.score(X_train, y_train)

grid.score(X_test, y_test)

grid.best_params_

imports = pd.DataFrame(
    {
        'feature': grid.best_estimator_.named_steps['tfidfvectorizer'].get_feature_names(),
        'coef': grid.best_estimator_.named_steps['multinomialnb'].coef_[0]
    }
)

imports.sort_values(by='coef', ascending=False).head(10).T

imports.sort_values(by='coef', ascending=False).tail(10).T

#### Results
Basic MultinomialNB scores similarly to many of the other models tested on accuracy.  Looking at the feature importances based off of the coefficients, we see many terms related to fitness and bodyweight fitness.  However, without preprocessing there are also many general terms with high influence likely due to their prevelence in the posts.

## Custom Preprocessing

def my_preprocessor(text):
    """Process text for use with vectorizer.
    
    Takes in a string and modifies it for use with count or tf-idf
    vectorizer.  It will set all characters to lowercase, remove and urls,
    remove the apostrophe from contractions, and replace non-alphanumeric
    characters with a space.
    
    Args:
        text (string): string to be processed
    
    Returns:
        string: processed text string
    """
    text = text.lower()
    text = re.sub('http\S+|www.\S+', '', text)
    text = text.replace("'", '')
    text = ''.join(w if w.isalnum() or w == ' ' else ' ' for w in text)
    
    return text

pipe2 = make_pipeline(TfidfVectorizer(preprocessor=my_preprocessor), 
                      StandardScaler(with_mean=False), MultinomialNB())
grid2 = GridSearchCV(pipe2, params, n_jobs=-1)

grid2.fit(X_train, y_train)

grid2.score(X_train, y_train)

grid2.score(X_test, y_test)

grid2.best_params_

## Custom Tokenizer
Testing model with both a lemmatizer and stemmer

def my_lemmatizer(text):
    """lemmatizer wrapper for use with vectorizer"""
    wnet = WordNetLemmatizer()
    return [wnet.lemmatize(w) for w in word_tokenize(text)]

def my_stemmer(text):
    """PoerterStemmer wrapper for use with vectorizer"""
    stemmer = PorterStemmer()
    return [stemmer.stem(w) for w in word_tokenize(text)]

# apply same pre-processing as posts
# inspiration from https://stackoverflow.com/questions/50155188/lemmatization-on-countvectorizer-doesnt-remove-stopwords
wnet = WordNetLemmatizer()
lem_stopwords = [wnet.lemmatize(w.replace("'",'')) for w in stopwords.words('english')]

stemmer = PorterStemmer()
stem_stopwords = [stemmer.stem(w.replace("'",'')) for w in stopwords.words('english')]

pipe3 = make_pipeline(
    TfidfVectorizer(preprocessor=my_preprocessor,
                    tokenizer=my_lemmatizer), 
    StandardScaler(with_mean=False), 
    MultinomialNB()
)

params3 = {
    'tfidfvectorizer__max_features': [None, 400, 800, 1000],
    'tfidfvectorizer__ngram_range': [(1,1), (1,2)],
    'tfidfvectorizer__stop_words': [None, lem_stopwords],
    'multinomialnb__alpha': [0.001, 0.01, 1]
}

grid3 = GridSearchCV(pipe3, params3, n_jobs=-1)

grid3.fit(X_train, y_train)

grid3.score(X_train, y_train)

grid3.score(X_test, y_test)

Best Params:  
{'multinomialnb__alpha': 0.001,  
 'tfidfvectorizer__max_features': 1000,  
 'tfidfvectorizer__ngram_range': (1, 2),  
 'tfidfvectorizer__stop_words': lem_stopwords}

imports3 = pd.DataFrame(
    {
        'feature': grid3.best_estimator_.named_steps['tfidfvectorizer'].get_feature_names(),
        'coef': grid3.best_estimator_.named_steps['multinomialnb'].coef_[0]
    }
)

imports3.sort_values(by='coef', ascending=False).head(10).T

imports3.sort_values(by='coef', ascending=False).tail(10).T

### Results
Adding in custom preprocessing and lemmetization seems to slightly decrease the models accuracy, but remove many of the general terms from the most important features based on the log probabilities of each feature.  Some of the features identified this way are terms that one would expect to be more associated with fitness or bodyweight fitness, but many still seem fairly generalized.
